backend-sync/requirements.txt
# Kuwait Social AI - Backend Requirements

# Core Framework
Flask==3.0.0  # Updated for security patches
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23  # Updated for security patches
alembic==1.13.1

# Redis & Caching
redis==5.0.1
celery==5.3.4  # Updated for bug fixes
cachetools==5.3.1

# Validation & Serialization
marshmallow==3.20.1
python-dotenv==1.0.0

# Authentication & Security
werkzeug==3.0.1  # Critical security update
PyJWT==2.8.0
zxcvbn==4.4.28  # For advanced password strength assessment

# AI & Content Generation
openai>=1.0.0
Pillow==10.1.0  # Security patches
deep-translator==1.11.4

# Data Processing
numpy==1.26.2

# File Processing
python-magic==0.4.27

# Security & Sanitization
bleach==6.1.0  # For secure HTML sanitization

# Social Media APIs
requests==2.31.0

# Date & Time (Kuwait timezone)
pytz==2023.3

# Utilities
click==8.1.7  # CLI framework for management commands

# Production Server (keep even though not imported)
gunicorn==21.2.0

# Note: Some packages that appear unused might be:
# 1. Required dependencies of other packages
# 2. Used dynamically or in configuration
# 3. Planned for future features
# See requirements-unused.txt for full list of potentially unused packages